requirements.txt
absl-py==2.3.1
affine==2.4.0
aiobotocore==2.25.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aioitertools==0.12.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.4.0
azure-core==1.36.0
azure-storage-blob==12.27.1
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.2.0
botocore==1.40.49
bounded-pool-executor==0.0.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-plugins==1.1.1.2
cligj==0.7.2
cloudpickle==3.1.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
curlify==2.2.1
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.8.0
earthaccess==0.15.1
email-validator==2.3.0
executing==2.2.1
fastapi==0.118.3
fastjsonschema==2.21.2
flatbuffers==25.9.23
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
gast==0.6.0
geopandas==1.1.1
google-pasta==0.2.0
grpcio==1.75.1
h11==0.16.0
h5py==3.14.0
harmony-py==1.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kaggle==1.7.4.5
keras==3.11.3
kiwisolver==1.4.9
lark==1.3.0
libclang==18.1.1
lightgbm==4.6.0
llvmlite==0.45.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
ml_dtypes==0.5.3
multidict==6.7.0
multimethod==2.0
namex==0.1.0
narwhals==2.8.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.7
notebook_shim==0.2.4
numba==0.62.1
numpy==2.2.6
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
plotly==6.3.1
pockets==0.9.1
pqdm==0.2.0
progressbar2==4.2.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
pyogrio==0.11.1
pyparsing==3.2.5
pyproj==3.7.2
python-cmr==0.13.0
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
python-json-logger==4.0.0
python-multipart==0.0.20
python-slugify==8.0.4
python-utils==3.9.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
rasterio==1.4.3
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rioxarray==0.19.0
rpds-py==0.27.1
s3fs==2025.9.0
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.9.0
shap==0.49.1
shapely==2.0.7
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
soupsieve==2.8
sphinxcontrib-napoleon==0.7
stack-data==0.6.3
starlette==0.48.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tinycss2==1.4.0
tinynetrc==1.3.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20251008
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.37.0
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
xarray==2025.10.1
xgboost==3.0.5
yarl==1.22.0
